Move methods to helper
