+++ lib/open_food_network/orders_and_fulfillments_report/customer_totals_report.rb
@@ -2,6 +2,8 @@
 module OpenFoodNetwork
   class OrdersAndFulfillmentsReport
     class CustomerTotalsReport
+      include OpenFoodNetwork::Reports::Helper
+
       REPORT_TYPE = "order_cycle_customer_totals".freeze
 
       attr_reader :context
@@ -193,45 +195,6 @@ module OpenFoodNetwork
         [{ variant: { product: :supplier },
            order: [:bill_address, :ship_address, :order_cycle, :adjustments, :payments,
                    :user, :distributor, shipments: { shipping_rates: :shipping_method }] }]
-      end
-
-      private
-
-      def admin_and_handling_total(order)
-        order.adjustments.select do |a|
-          a.eligible && a.originator_type == 'EnterpriseFee' && a.source_type != 'Spree::LineItem'
-        end.map(&:amount).sum.to_f
-      end
-
-      def ship_total(order)
-        order.adjustments.select{ |a| a.originator_type == 'Spree::ShippingMethod' }.
-          map(&:amount).sum.to_f
-      end
-
-      def payment_fee(order)
-        order.adjustments.select{ |a| a.originator_type == 'Spree::PaymentMethod' }.
-          map(&:amount).sum.to_f
-      end
-
-      def amount_with_adjustments(line_items)
-        line_items.map do |line_item|
-          price_with_adjustments(line_item) * line_item.quantity
-        end.sum
-      end
-
-      def price_with_adjustments(line_item)
-        return 0 if line_item.quantity == 0
-        (line_item.price + line_item_adjustments(line_item).map(&:amount).sum / line_item.quantity).
-          round(2)
-      end
-
-      def line_item_adjustments(line_item)
-        line_item.order.adjustments.select{ |a| a.source_id == line_item.id }
-      end
-
-      def shipping_method(line_items)
-        line_items.first.order.shipments.first.
-          andand.shipping_rates.andand.first.andand.shipping_method
       end
     end
   end
